@@ -50,7 +50,10 @@ class ZikrViewerPageBuilder extends StatelessWidget {
                     ? S.of(context).done
                     : "${dbContent.count}".toArabicNumber(),
                 style: TextStyle(
-                  color: Theme.of(context).colorScheme.primary.withOpacity(.02),
+                  color: Theme.of(context)
+                      .colorScheme
+                      .primary
+                      .withValues(alpha: 0.02),
                   fontSize: 250,
                   fontWeight: FontWeight.bold,
                 ),
